Extract getAllEvents to a new file
